Split toast into toast and stack
